requirements.txt
flask>=3.0
requests>=2.31
python-dateutil>=2.8
colorama>=0.4
beautifulsoup4>=4.12
lxml>=5.0
